requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
backoff==2.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
emoji==2.14.1
eval_type_backport==0.2.2
filetype==1.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
jsonpath-python==1.0.6
langdetect==1.0.9
lxml==5.3.1
marshmallow==3.26.1
mistralai==1.5.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
psutil==7.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pypdf==5.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.1
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tabulate==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.25
unstructured-client==0.31.1
unstructured-ingest==0.5.13
urllib3==2.3.0
webencodings==0.5.1
wrapt==1.17.2
zipp==3.21.0
